+++ src/main/kotlin/de/claudioaltamura/kotlin_spring_boot_todo/repository/TodoRepository.kt
@@ -4,7 +4,7 @@ import de.claudioaltamura.kotlin_spring_boot_todo.entity.TodoEntity
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.CrudRepository
 
-interface TodoRepository : CrudRepository<TodoEntity, Int> {
+interface TodoRepository : CrudRepository<TodoEntity, Long> {
 
     @Query(value = "SELECT * FROM TODOS WHERE title like %?1%", nativeQuery = true)
     fun findByTitle(title : String) : List<TodoEntity>
